feat: you can now ignore votes from players who are not online on the bukkit server.
I may have left out something to be modified so that the vote is ignored, I know. It is the first time that all this project so I do not quite understand it yet, if I find something else that needs to be modified for this option to work properly or something that generates abnormalities when ignoring the votes of disconnected people, I will fix it.